requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Flask-Compress==1.13
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
email-validator==2.1.0
APScheduler==3.10.4